requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-core
langchain-openai
langchain-pinecone
pinecone-client
pydantic
slowapi
langchain-community
pypdf
watchdog
cryptography>=42.0.0
mangum